requirements.txt
Flask==2.2.5
keras==2.11.0
numpy==1.21.6
Pillow==9.5.0
tensorflow==2.11.0
